Fixed misplaced error log message and moved FinalizeRollappPackets to hooks as well.
